requirements.txt
aioredis==1.2.0
amqp==2.6.0
aniso8601==7.0.0
argh==0.26.2
asgiref==3.2.10
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
billiard==3.6.3.0
celery==4.4.6
certifi==2019.9.11
cffi==1.12.3
channels==2.3.1
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.4.1
defusedxml==0.6.0
Django==3.0
django-allauth==0.41.0
django-cleanup==5.0.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-jwt==0.3.4
django-model-utils==4.1.1
django-rest-auth==0.9.5
django-split-settings==1.1.0
djangorestframework==3.10.3
docxtpl==0.16.7
et-xmlfile==1.0.1
filetype==1.2.0
future==0.18.2
graphene==2.1.8
graphene-django==2.10.0
graphene-django-optimizer==0.6.2
graphene-django-plus==2.1
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==1.0.0
html5lib==1.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
inflect==5.0.2
jdcal==1.4.1
kombu==4.6.11
Markdown==3.1.1
msgpack==0.6.2
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
pathtools==0.1.2
pdfkit==1.0.0
Pillow==6.2.1
promise==2.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==2.7.0
pyOpenSSL==19.0.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.2
reportlab==3.5.51
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
selenium==3.141.0
sendgrid==3.6.5
sendgrid-django==4.2.0
service-identity==21.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.5
vine==1.3.0
watchdog==0.10.3
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.4
zipp==3.1.0
zope.interface==4.7.2
